code review changes: use NOMINMAX hack, audio backend android only, move windows time defs to time_defs.hpp
