+++ CMakeLists.txt
@@ -207,12 +207,9 @@ if(WIN32)
     find_package(FLAC REQUIRED)
     find_package(Ogg REQUIRED)
     find_package(Vorbis REQUIRED)
-    find_package(Opus)    
-    if(OPUS_FOUND)
-        add_definitions("-DHAS_OPUS")
-    endif()
+    find_package(Opus REQUIRED)    
     add_definitions(-DNTDDI_VERSION=0x06020000 -D_WIN32_WINNT=0x0602 -DWINVER=0x0602 -DWINDOWS -DWIN32_LEAN_AND_MEAN -DUNICODE -D_UNICODE -D_CRT_SECURE_NO_WARNINGS )
-    add_definitions(-DHAS_OGG -DHAS_VORBIS -DHAS_FLAC -DHAS_VORBIS_ENC -DHAS_WASAPI)
+    add_definitions(-DHAS_OGG -DHAS_VORBIS -DHAS_FLAC -DHAS_VORBIS_ENC -DHAS_WASAPI -DHAS_OPUS)
 endif()
 
 add_subdirectory(common)
