+++ common/message/message.hpp
@@ -30,31 +30,6 @@
 #endif
 #include <vector>
 
-#ifdef WINDOWS // Implementation from http://stackoverflow.com/a/26085827/2510022
-#include <Windows.h>
-#include <stdint.h>
-#include <winsock2.h>
-
-inline static int gettimeofday(struct timeval* tp, struct timezone* tzp)
-{
-    // Note: some broken versions only have 8 trailing zero's, the correct epoch has 9 trailing zero's
-    static const uint64_t EPOCH = ((uint64_t)116444736000000000ULL);
-
-    SYSTEMTIME system_time;
-    FILETIME file_time;
-    uint64_t time;
-
-    GetSystemTime(&system_time);
-    SystemTimeToFileTime(&system_time, &file_time);
-    time = ((uint64_t)file_time.dwLowDateTime);
-    time += ((uint64_t)file_time.dwHighDateTime) << 32;
-
-    tp->tv_sec = (long)((time - EPOCH) / 10000000L);
-    tp->tv_usec = (long)(system_time.wMilliseconds * 1000);
-    return 0;
-}
-#endif
-
 /*
 template<typename CharT, typename TraitsT = std::char_traits<CharT> >
 class vectorwrapbuf : public std::basic_streambuf<CharT, TraitsT>
@@ -96,11 +71,7 @@ struct tv
     tv()
     {
         timeval t;
-#ifdef WINDOWS
-        gettimeofday(&t, NULL);
-#else
         chronos::steadytimeofday(&t);
-#endif
         sec = t.tv_sec;
         usec = t.tv_usec;
     }
